requirements.txt
Flask==3.0.3
numpy==1.26.4
pandas==2.2.2
PyAudio==0.2.14
pydub==0.25.1
requests==2.32.3
scipy
sounddevice==0.5.0
SpeechRecognition==3.10.4
torch==2.4.0
transformers==4.44.0
openai-whisper